+++ CHANGELOG.md
@@ -5,6 +5,36 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.67.0] - 2023-06-01 03:05:10
+
+### Added
+
+- Quebec Senior Assistance Tax Credit.
+
+## [0.66.0] - 2023-05-30 16:16:25
+
+### Added
+
+- Manitoba Pension amount credit.
+
+## [0.65.0] - 2023-05-28 20:47:19
+
+### Added
+
+- New Brunswick age amount credit.
+
+## [0.64.0] - 2023-05-27 23:28:36
+
+### Added
+
+- Quebec childcare expenses tax credit.
+
+## [0.63.0] - 2023-05-27 15:38:10
+
+### Added
+
+- Python 3.10 support.
+
 ## [0.62.0] - 2023-05-25 22:46:06
 
 ### Added
@@ -504,6 +534,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[0.67.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.66.0...0.67.0
+[0.66.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.65.0...0.66.0
+[0.65.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.64.0...0.65.0
+[0.64.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.63.0...0.64.0
+[0.63.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.62.0...0.63.0
 [0.62.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.61.0...0.62.0
 [0.61.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.60.0...0.61.0
 [0.60.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.59.0...0.60.0